backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
channels==4.1.0
daphne==4.1.2
python-dotenv==1.2.1
dj-database-url==3.0.1
psycopg[binary]==3.2.9
requests==2.32.5
